% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{RMWreg_MCMC}
\alias{RMWreg_MCMC}
\title{MCMC sampler to fit a RMW AFT regression model}
\usage{
RMWreg_MCMC(N, Thin, Burn, Time, Event, DesignMat, Mixing = "None",
  BaseModel = "Weibull", PriorCV = "Pareto", PriorMeanCV = 1.5,
  Hyp1Gam = 1, Hyp2Gam = 1, ...)
}
\arguments{
\item{N}{Total number of iterations for the MCMC sampler. Use \code{N>=max(4,Thin)}, \code{N} being a multiple of \code{Thin}.}

\item{Thin}{Thining period for the MCMC sampler. Use \code{Thin>=2}.}

\item{Burn}{Burn-in period for the MCMC sampler. Use \code{Burn>=1}, \code{Burn<N}, \code{Burn} being a multiple of \code{Thin}.}

\item{Time}{Vector of length \code{n} containing survival times}

\item{Event}{Vector of length \code{n} containing event indicators (\code{TRUE} / \code{1} is the event is observed, \code{FALSE} / \code{0} is the event is Eventored)}

\item{Mixing}{Mixing distribution assigned to the (frailty) random effects. Possible values are
\code{"None", "Exponential", "Gamma", "InvGamma", "InvGauss", "LogNormal"}}

\item{BaseModel}{If \code{BaseModel = "Weibull"}, a RMW regression is used. If \code{BaseModel = "Exponential"}, a RME regression is used.}

\item{PriorCV}{Type of prior assigned to the coefficient of variation of the survival times.
Possible values are \code{"Pareto", "TruncExp"}}

\item{PriorMeanCV}{Ellicited prior mean of the coefficient of variation (\code{PriorMeanCV > 1}).
If \code{Mixing = "InvGamma"}, \code{PriorMeanCV} must be below \code{sqrt{3}}.
If \code{Mixing = "InvGauss"}, \code{PriorMeanCV} must be below \code{sqrt{5}}.
Default: \code{PriorMeanCV = 1.5}}

\item{Hyp1Gam}{Shape hyper-parameter for the Gamma(\code{Hyp1Gam}, \code{Hyp2Gam}) assigned to \code{gam}.}

\item{Hyp2Gam}{Rate hyper-parameter for the Gamma(\code{Hyp1Gam}, \code{Hyp2Gam}) assigned to \code{gam}.}

\item{...}{Optional parameters.
\describe{

\item{\code{AR}}{Optimal acceptance rate for adaptive Metropolis Hastings updates.
It must be a positive number between 0 and 1.
Default (and recommended): \code{ar = 0.44}}.
\item{\code{StopAdapt}}{Iteration at which adaptive proposals are not longer adapted.
Use \code{stopAdapt>=1}. Default: \code{StopAdapt = Burn}.}
\item{\code{StoreChains}}{If \code{StoreChains = TRUE}, MCMC chains of each parameter are stored in separate .txt files.
(\code{RunName} argument used for file names). Default: \code{StoreChains = FALSE}.}
\item{\code{StoreAdapt}}{If \code{StoreAdapt = TRUE}, trajectory of adaptive proposal variances (log scale)
for each parameter are stored in separate .txt files. (\code{RunName} argument used for file names).
Default: \code{StoreAdapt = FALSE}.}
\item{\code{StoreDir}}{Directory where MCMC chain will be stored (only required if \code{Store = TRUE}).
Default: \code{StoreDir = getwd()}.}
\item{\code{RunName}}{Run-name to be used when storing chains and/or adaptive proposal variances in .txt files.}
\item{\code{PrintProgress}}{If \code{PrintProgress = TRUE}, intermediate output is displayed in the console. }
}}
}
\value{
A \code{list} containing MCMC draws for all parameters.
}
\description{
MCMC sampler to fit a RMW AFT regression model
}
\examples{

library(KMsurv)
data(alloauto)
n=dim(alloauto)[1]; k=2
Intercept=rep(1,times=n); x1=alloauto$type-1
DesignMat=cbind(Intercept,x1); rm(Intercept)
Time=alloauto$time; Event=alloauto$delta

Chain <- RMWreg_MCMC(N = 100, Thin = 2, Burn = 50,
                     Time, Event, DesignMat,
                     Mixing = "None", BaseModel = "Weibull",
                     PriorCV = "Pareto", PriorMeanCV = 1.5,
                     Hyp1Gam = 1, Hyp2Gam = 1)

}
\author{
Catalina A. Vallejos \email{cvallejos@turing.ac.uk}
}

