% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{RMWreg_PredictCDF}
\alias{RMWreg_PredictCDF}
\title{CDF prediction based on a pre-computed MCMC chain.}
\usage{
RMWreg_PredictCDF(Chain, Time, x, Mixing = "None", BaseModel = "Weibull")
}
\arguments{
\item{Chain}{List containing the pre-computed MCMC sample (as generated by the function \code{RMWreg_MCMC}).}

\item{Time}{Time threshold \eqn{t} at which the predictive CDF is estimated.}

\item{x}{Vector of covariate information for the new subject (must be in the same order as in the
design matrix used to fit the model, including a '1' on the first position to represent the intercept).}

\item{Mixing}{Mixing distribution assigned to the (frailty) random effects. Possible values are
\code{"None", "Exponential", "Gamma", "InvGamma", "InvGauss", "LogNormal"}}

\item{BaseModel}{If \code{BaseModel = "Weibull"}, a RMW regression is used. If \code{BaseModel = "Exponential"}, a RME regression is used.}
}
\value{
A Monte Carlo estimate of the predictive CDF.
}
\description{
This functions returns a Monte Carlo estimate of the predictive CDF \eqn{P(T < t | D)}, based on a (pre-computed)
posterior sample of all model parameters of the associated RMW regression model (obtained based on a dataset \eqn{D}).
}
\examples{

library(KMsurv)
data(alloauto)
n=dim(alloauto)[1]; k=2
Intercept=rep(1,times=n); x1=alloauto$type-1
DesignMat=cbind(Intercept,x1); rm(Intercept)
Time=alloauto$time; Event=alloauto$delta

Chain <- RMWreg_MCMC(N = 100, Thin = 2, Burn = 50,
                     Time, Event, DesignMat,
                     Mixing = "None", BaseModel = "Weibull",
                     PriorCV = "Pareto", PriorMeanCV = 1.5,
                     Hyp1Gam = 1, Hyp2Gam = 1)

# Predictive CDF at time = 5, for a given covariate configuration
RMWreg_PredictCDF(Chain, Time = 5, x = c(1,1), Mixing = "None", BaseModel = "Weibull")

}
\author{
Catalina A. Vallejos \email{cvallejos@turing.ac.uk}
}

